+++ src/lang/en.js
@@ -40,7 +40,10 @@ export default {
     select_district: 'Select District/City',
     select_sub_district: 'Select Sub District',
     select_village: 'Select Village',
-    must_be_filled: 'Must be filled'
+    must_be_filled: 'Must be filled',
+    kit_received: 'Kit Received',
+    kit_available: 'Kit Available',
+    used_kit: 'Used Kit'
   },
   success: {
     login_success: 'Login Success',
