+++ src/store/modules/recipient/actions.js
@@ -19,5 +19,14 @@ export default {
     } catch (error) {
       return error.response
     }
+  },
+  async getRecipient({ commit }, params) {
+    try {
+      const response = await fetchList('/api/v1/recipients', 'GET', params)
+      commit('SET_RECIPIENT', response.data.data[0])
+      return response
+    } catch (error) {
+      return error.response
+    }
   }
 }
